# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/students": {
      "get": {
        "operationId": "StudentsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "학생 원부 목록 — 관리자 전체 / 강사는 담당 학생만+안전 필드(P0-5) [전 직원]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "post": {
        "operationId": "RegistrationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterStudentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 aggregate 생성 — /students/registrations와 동일한 원자 command",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}/family-relations": {
      "get": {
        "operationId": "StudentsController_findFamilyRelations",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "학생의 등록 가족 관계 목록 조회 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "post": {
        "operationId": "StudentsController_createFamilyRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentFamilyRelationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생과 기존 학생의 가족 관계 생성 및 이력 기록 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}/family": {
      "get": {
        "operationId": "StudentsController_findFamilyAggregate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "가족(형제·자매) 조인 aggregate — 구성원별 학생·보호자·활성 수강·상담 이력·공유 보호자(테이블 조인 파생, 단일 진실원) [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}/family-relations/{relationId}": {
      "patch": {
        "operationId": "StudentsController_updateFamilyRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "relationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentFamilyRelationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 가족 관계 수정 및 이력 기록 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentsController_removeFamilyRelation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "relationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "학생 가족 관계 soft delete 및 이력 기록 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}/academic-histories": {
      "get": {
        "operationId": "StudentsController_findAcademicHistories",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "학생 학교·학년 변경 timeline 조회 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "post": {
        "operationId": "StudentsController_createAcademicHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentAcademicHistoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 학교·학년 이력 생성 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}/academic-histories/{historyId}": {
      "patch": {
        "operationId": "StudentsController_updateAcademicHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "historyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentAcademicHistoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 학교·학년 이력 수정 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentsController_removeAcademicHistory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          },
          {
            "name": "historyId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "학생 학교·학년 이력 soft delete [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/students/{id}": {
      "get": {
        "operationId": "StudentsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 단건 — 관리자 full / 강사는 담당 학생만+안전 필드, 밖은 403(P0-5) [전 직원]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "delete": {
        "operationId": "StudentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 soft delete 및 관련 활성 관계 정리(재인증 필수) [매니저 이상]. cookie 세션은 reauth 후 10분 내만 허용(403 SUDO_REQUIRED).",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      },
      "patch": {
        "operationId": "RegistrationsController_updateStudent",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 1,
              "maximum": 2147483647
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "학생 기본 프로필 수정 및 audit 기록 [매니저 이상]",
        "tags": [
          "students"
        ],
        "security": [
          {
            "bearer": []
          },
          {
            "access_token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "TACO ERP API",
    "description": "TnAcademy 백오피스 운영 API. Postgres 영속 저장소와 역할 기반 접근 제어를 사용하는 실제 서비스 계약.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/",
      "description": "현재 호스트"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "access_token": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
...
      "CreateStudentFamilyRelationDto": {
        "type": "object",
        "properties": {
          "relatedStudentId": {
            "type": "number",
            "minimum": 1
          },
          "relationType": {
            "type": "string",
            "enum": [
              "sibling",
              "other"
            ]
          },
          "relationLabel": {
            "type": "string",
            "maxLength": 50
          },
          "linkGuardians": {
            "type": "boolean",
            "description": "[TBO-30G] true면 같은 tx에서 두 학생의 보호자를 관계 행(join)으로 합집합 연결 —\n 보호자 원부 복사 0, 신규 링크는 비대표·비납부(기존 대표 불변), 중복 연결은 건너뜀."
          }
        },
        "required": [
          "relatedStudentId",
          "relationType"
        ]
      },
      "UpdateStudentFamilyRelationDto": {
        "type": "object",
        "properties": {
          "relationType": {
            "type": "string",
            "enum": [
              "sibling",
              "other"
            ]
          },
          "relationLabel": {
            "type": "string",
            "maxLength": 50
          }
        }
      },
      "CreateStudentAcademicHistoryDto": {
        "type": "object",
        "properties": {
          "grade": {
            "type": "number",
            "minimum": 0,
            "maximum": 13
          },
          "schoolName": {
            "type": "string",
            "maxLength": 100
          },
          "startedOn": {
            "type": "string"
          },
          "endedOn": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "grade",
          "schoolName",
          "startedOn"
        ]
      },
      "UpdateStudentAcademicHistoryDto": {
        "type": "object",
        "properties": {
          "grade": {
            "type": "number",
            "minimum": 0,
            "maximum": 13
          },
          "schoolName": {
            "type": "string",
            "maxLength": 100
          },
          "startedOn": {
            "type": "string"
          },
          "endedOn": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "StudentInterestDto": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "number",
            "minimum": 1,
            "description": "기존 course 선택. customLabel과 정확히 하나만 허용"
          },
          "customLabel": {
            "type": "string",
            "maxLength": 120,
            "description": "카탈로그 외 희망 수업. courseId와 정확히 하나만 허용"
          },
          "priority": {
            "type": "number",
            "minimum": 1,
            "description": "1부터 시작하는 연속 우선순위"
          }
        },
        "required": [
          "priority"
        ]
      },
...
      "CreateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 20
          },
          "englishName": {
            "type": "string",
            "maxLength": 50
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other",
              "undisclosed"
            ]
          },
          "birthDate": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "grade": {
            "type": "number",
            "minimum": 0,
            "maximum": 13
          },
          "schoolName": {
            "type": "string",
            "maxLength": 100
          },
          "residenceType": {
            "type": "string",
            "enum": [
              "domestic",
              "overseas"
            ]
          },
          "address": {
            "type": "string",
            "maxLength": 100
          },
          "addressDetail": {
            "type": "string",
            "maxLength": 100
          },
          "kakaoId": {
            "type": "string",
            "maxLength": 100
          },
          "counselTopic": {
            "type": "string",
            "maxLength": 1000
          },
          "status": {
            "type": "string",
            "enum": [
              "enrolled",
              "on_leave",
              "withdrawn",
              "registration_lost",
              "new_inquiry"
            ]
          },
          "country": {
            "type": "string",
            "pattern": "^[A-Z]{2}$"
          },
          "memo": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name",
          "birthDate",
          "grade"
        ]
      },
      "RegistrationGuardianDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 20,
            "pattern": "\\S",
            "example": "김학부모",
            "description": "보호자 이름"
          },
          "phone": {
            "type": "string",
            "maxLength": 20,
            "example": "010-1234-5678",
            "description": "연락처 — 같은 번호는 기존 보호자에 연결(upsert-or-link)"
          },
          "relation": {
            "type": "string",
            "maxLength": 20,
            "example": "모",
            "description": "학생과의 관계(모/부/조모 등 자유 문자열)"
          },
          "isPayer": {
            "type": "boolean",
            "description": "납부자 여부(기본 true)"
          },
          "isPrimary": {
            "type": "boolean",
            "description": "대표 보호자 여부(기본 true — 기존 대표는 자동 강등)"
          }
        },
        "required": [
          "name"
        ]
      },
      "RegisterStudentDto": {
        "type": "object",
        "properties": {
          "student": {
            "description": "학생 정보(기존 POST /students와 동일 필드)",
            "allOf": [
              {
                "$ref": "#/components/schemas/CreateStudentDto"
              }
            ]
          },
          "interests": {
            "minItems": 2,
            "maxItems": 20,
            "description": "희망 수업 2개 이상",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StudentInterestDto"
            }
          },
          "guardians": {
            "maxItems": 10,
            "description": "보호자 0~10명",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RegistrationGuardianDto"
            }
          },
          "guardian": {
            "deprecated": true,
            "description": "v0.2.5 호환 단건 보호자. guardians와 동시 사용 금지",
            "allOf": [
              {
                "$ref": "#/components/schemas/RegistrationGuardianDto"
              }
            ]
          },
          "courseId": {
            "type": "number",
            "minimum": 1,
            "example": 10,
            "description": "수강 코스 — 있으면 enrollment를 같은 tx로 생성"
          }
        },
        "required": [
          "student",
          "interests"
        ]
      },
      "UpdateStudentDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 20
          },
          "englishName": {
            "type": "string",
            "maxLength": 50
          },
          "gender": {
            "enum": [
              "male",
              "female",
              "other",
              "undisclosed"
            ],
            "type": "string"
          },
          "birthDate": {
            "type": "string"
          },
          "grade": {
            "type": "number",
            "minimum": 0,
            "maximum": 13
          },
          "schoolName": {
            "type": "string",
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "maxLength": 20
          },
          "country": {
            "type": "string",
            "pattern": "^[A-Z]{2}$"
          },
          "residenceType": {
            "enum": [
              "domestic",
              "overseas"
            ],
            "type": "string"
          },
          "address": {
            "type": "string",
            "maxLength": 100
          },
          "addressDetail": {
            "type": "string",
            "maxLength": 100
          },
          "kakaoId": {
            "type": "string",
            "maxLength": 100
          },
          "counselTopic": {
            "type": "string",
            "maxLength": 1000
          },
          "status": {
            "enum": [
              "enrolled",
              "on_leave",
              "withdrawn",
              "registration_lost",
              "new_inquiry"
            ],
            "type": "string"
          },
          "memo": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "access_token": []
    }
  ]
}